Restructuring to see if this in shaa Allah can make it so that Pkg.add works
